+++ toolz/__init__.py
@@ -4,8 +4,6 @@ from .functoolz import *
 
 from .dicttoolz import *
 
-from .objtoolz import *
-
 from .recipes import *
 
 from .compatibility import map, filter
