+++ toolz/sandbox/__init__.py
@@ -1,1 +1,3 @@
 from .parallel import fold
+
+from .objtoolz import *
