+++ toolz/sandbox/objtoolz.py
